[UI] Исправлено отображение заявок в друзья Resolve #1209
